Rename Grid module from GridDisplay
