+++ crosswrd-app/src/EditLights.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { Container } from "react-bootstrap";
 
 import { Grid, CellMap } from "./Lights";
-import { displayGrid, renderCells } from "./GridDisplay";
+import { displayGrid, renderCells } from "./Grid";
 import { WrappedRow, StateSetter, ToggleButton } from "./Helpers";
 import { ExportLights } from "./ExportLights";
